Configure cross-platform binary releases
We set up TravisCI and AppVeyor to upload binary releases for
Linux (static!), MacOS X and Windows (MSVC). In order to make this work,
I had to replace two C++ depdendencies:

- `uchardet` replaced with `chardet`
- `cld2` replaced with `whatlang`

Trying to cross-compile C++ dependencies for `musl-libc` with `cmake`
was an exercise in misery, and I'm happy that pure Rust versions of both
libraries exist.