Vobsub: work harder to get `needed` length
The `needed` length was not available in all cases which made it
useless in some situations in chunk mode.

This commit attempts to guestimate the `needed` length so that
the next chunk is always large enough for the next iteration to
at least give a new `needed` length which will allow progressing
in the stream.