Problem: tests fail on Windows
Also, if a non-exe file maching binary is present in the
build directory, it'll get picked up but there'll be an error
saying that this binary is not a valid Win32 binary.

Solution: adjust tests to accommodate for Windows' ecosystem
and make the binary name automatically assume .exe extension
when compiled on Windows.